fix: thread pending_click into AppRuntime::Builtin; File Explorer click dispatch (stint 0469)
App::ui now takes pending_click so builtin apps can see synthetic
node-targeted clicks the same way WASM/Python apps already do.
INTERACTIVE_ROLES gains checkbox; PaneClickTarget::Node widens u32->u64
since AccessKit-derived builtin node ids are full-width egui::Id hashes,
not small WASM arena indices. File Explorer's "Show hidden files"
checkbox is the reference implementation via identity-match on its own
egui::Id, mirroring node_click_matches on the WASM/Python path.

+++ src/host/wasm_render.rs
@@ -229,7 +229,7 @@ fn cross_align(a: Alignment) -> egui::Align {
 fn node_click_matches(pending_click: Option<crate::host::pane::PendingPaneClick>, id: u32) -> bool {
     matches!(
         pending_click.map(|c| c.target),
-        Some(crate::host::pane::PaneClickTarget::Node(n)) if n == id
+        Some(crate::host::pane::PaneClickTarget::Node(n)) if n == u64::from(id)
     )
 }
